event-generator/requirements.txt
kafka-python==2.0.2
avro-python3==1.10.2
faker==20.1.0
python-dotenv==1.0.0